XDP option_open fix

+++ apisrc/EkaFhPlrRecovery.c
@@ -136,7 +136,7 @@ static EkaOpResult sendRetransmissionRequest(EkaFhPlrGr* gr, int sock, uint32_t 
 
   EKA_LOG("Sending RetransmissionRequest: BeginSeqNum=%u,EndSeqNum=%u,SourceID=\'%.*s\',ProductID=%u,ChannelID=%u",
 	  msg->BeginSeqNum,msg->EndSeqNum,
-          sizeof(msg->SourceID),msg->SourceID,msg->ProductID,msg->ChannelID);
+          (int)sizeof(msg->SourceID),msg->SourceID,msg->ProductID,msg->ChannelID);
   int rc = send(sock,pkt,pktHdr->pktSize,0);
   if (rc <= 0) {
     EKA_WARN("Tcp send of msg size %d to sock %d returned rc = %d",
